requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
pytest
requests
pre-commit
python-dotenv
aiofiles
pillow
pytesseract